Module serde_codegen::attr

Structs

ContainerAttrs
Represents container (e.g. struct) attribute information

FieldAttrs
Represents field attribute information

Name

Respanner
This syntax folder rewrites tokens to say their spans are coming from a macro context.

VariantAttrs
Represents variant attribute information

Functions

get_renames

get_serde_meta_items

get_str_from_lit

get_struct_field_attrs
Extract out the #[serde(...)] attributes from a struct field.

parse_lit_into_path

wrap_default
This function wraps the expression in #[serde(default="...")] in a function to prevent it from accessing the internal Deserialize state.

wrap_serialize_with
This function wraps the expression in #[serde(serialize_with="...")] in a trait to prevent it from accessing the internal Serialize state.

wrap_skip_serializing
This function wraps the expression in #[serde(skip_serializing_if="...")] in a trait to prevent it from accessing the internal Serialize state.